% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beqi2-pkg.R
\docType{package}
\name{BEQI2-pkg}
\alias{BEQI}
\alias{BEQI2-pkg}
\alias{BEQI2-pkg-package}
\title{Marine Benthic Ecosystem Quality Index}
\description{
Tools for analysing marine benthic ecosystems.
}
\keyword{package}

