\name{entropy}
\alias{entropy}
\title{Shannon's Entropy}
\usage{
entropy(taxon, count)
}
\arguments{
  \item{taxon}{character vector with taxa}

  \item{count}{integer vector with counts}
}
\value{
Shannon's entropy
}
\description{
Compute entropy according to Shannon (1948)
}
\references{
Shannon, C. E., 1948. A Mathematical Theory of
Communication.  Bell System Technical Journal 27: 379-423.
}

