% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{margalef}
\alias{margalef}
\title{Margalef Index of Diversity}
\usage{
margalef(taxon, count)
}
\arguments{
\item{taxon}{character vector with taxa}

\item{count}{integer vector with counts}
}
\value{
Margalef diversity index (numeric vector of length 1)
}
\description{
Margalef Index of Diversity is given by
 \deqn{D = \frac{S-1}{\ln(N)}}{D = (S-1)/ln(N)}
}
\details{
For \eqn{N=1}{N=1}, the index is set to 0.
}

