% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{readSettings}
\alias{readSettings}
\title{Read BEQI Settings File}
\usage{
readSettings(filename)
}
\arguments{
\item{filename}{name of BEQI input file (\code{character})}
}
\description{
This function reads BEQI settings files (JSON)
}
\details{
The function performs the following tasks:
	\itemize{
		\item{checks the existence of \code{filename};}
 	\item{reads JSON file while ignoring C-style comments;}
     \item{checks avaiability of required keys in the JSON-file}
 	\item{checks values in JSON-file}
 }
}

