% Generated by roxygen2 (4.0.1): do not edit by hand
\name{speciesRichness}
\alias{speciesRichness}
\title{Species Richness}
\usage{
speciesRichness(taxon, count)
}
\arguments{
\item{taxon}{character vector with taxa}

\item{count}{integer vector with counts}
}
\value{
integer vector with abundance per taxon.
}
\description{
Compute species richness
}

