\name{isAzoic}
\alias{isAzoic}
\title{Test on Azoic Samples}
\usage{
isAzoic(x)
}
\arguments{
  \item{x}{character vector containing taxa}
}
\value{
logical vector, with elements \code{TRUE} for azoic
samples, and \code{FALSE} otherwise.
}
\description{
Case-insensitive test for taxa starting with 'azoi'
}

