% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{isAzoic}
\alias{isAzoic}
\title{Test for Azoic Samples}
\usage{
isAzoic(x)
}
\arguments{
\item{x}{character vector containing taxa}
}
\value{
logical vector, with elements \code{TRUE} for azoic samples,
     and \code{FALSE} otherwise.
}
\description{
Case-insensitive test for taxa starting with 'azoi'
}

