% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{speciesRichness}
\alias{speciesRichness}
\title{Species Richness}
\usage{
speciesRichness(taxon, count)
}
\arguments{
\item{taxon}{character vector with taxa}

\item{count}{integer vector with counts}
}
\value{
species richness (integer vector of length 1)
}
\description{
Species richness (\eqn{S}{S}) is defined as the number of taxa
 (lowest identification level possible) per sampling unit
 (data pool or box core sample).
}

